build: decouple box library from Tarantool binary
Currently it is impossible to use box library without Tarantool binary
because parts of box depend on main.{h, cc}: move the dependencies to
box.{h, cc}.

Needed for #8147

NO_CHANGELOG=refactoring
NO_DOC=refactoring
NO_TEST=refactoring
