+++ src/main.h
@@ -37,38 +37,11 @@
 extern "C" {
 #endif /* defined(__cplusplus) */
 
-double tarantool_uptime(void);
-
-/**
- * Path to Tarantool binary of the process.
- */
-extern char tarantool_path[];
-
-/**
- * Application start time on monotonic clocks.
- */
-extern long tarantool_start_time;
-
 void
 tarantool_exit(int);
 
 void
 load_cfg(void);
-
-/* The type of sigint callback's pointer. */
-struct ev_loop;
-struct ev_signal;
-typedef void
-(*sigint_cb_t)(struct ev_loop *loop, struct ev_signal *w, int revents);
-
-/*
- * This function is needed for setting the new sigint callback.
- * Returns the pointer to the old callback function.
- * The main scenario is to replace the current callback
- * and having an opportunity to set the old one back.
- */
-sigint_cb_t
-set_sigint_cb(sigint_cb_t new_sigint_cb);
 
 #if defined(__cplusplus)
 } /* extern "C" */
